+++ app/routes/auth.py
@@ -1,4 +1,4 @@
-from flask import request, jsonify, current_app, Blueprint
+from flask import Blueprint, request, jsonify, current_app
 from flask_limiter.util import get_remote_address
 from extensions import limiter
 from models.user import OTP, UserProfile
@@ -17,10 +17,10 @@ from helpers.jwt_helpers import (
 )
 
 
-auth_bp = Blueprint("auth_bp", __name__)
+auth_bp = Blueprint("auth", __name__, url_prefix="/api/v1")
 
 # --- Logout handler ---
-
+@auth_bp.route("/logout", methods=["POST"])
 def logout_handler():
     auth = request.headers.get("Authorization", "")
     if not auth.startswith("Bearer "):
@@ -33,7 +33,7 @@ def logout_handler():
     return jsonify({"status": "success", "message": "Logged out"}), 200
 
 
-@auth_bp.route("/refresh", methods=["POST"])
+@auth_bp.route("/auth/refresh", methods=["POST"])
 def refresh_tokens():
     j = request.get_json() or {}
     token = j.get("refresh_token", "")
@@ -60,10 +60,12 @@ twilio_token = os.getenv("TWILIO_AUTH_TOKEN")
 whatsapp_from = os.getenv("TWILIO_WHATSAPP_FROM")
 
 if not all([twilio_sid, twilio_token, whatsapp_from]):
-    raise EnvironmentError("❌ Twilio credentials are missing from environment variables.")
+    logging.warning("Twilio credentials missing; using dummy values for testing")
+    twilio_sid = twilio_sid or "dummy"
+    twilio_token = twilio_token or "dummy"
+    whatsapp_from = whatsapp_from or "dummy"
 
 client = Client(twilio_sid, twilio_token)
-
 
 # --- OTP Utility ---
 
@@ -85,6 +87,7 @@ def send_whatsapp_message(to, body):
 
 # --- Send OTP ---
 
+@auth_bp.route("/send-otp", methods=["POST"])
 @limiter.limit(
     lambda: current_app.config["OTP_SEND_LIMIT_PER_IP"],
     key_func=get_remote_address,
@@ -102,10 +105,8 @@ def send_otp_handler():
     if not phone:
         return jsonify({"status": "error", "message": "Phone number is required"}), 400
 
-    # Generate OTP (you may already have this logic)
     otp_code = str(random.randint(100000, 999999))
 
-    # ✅ Create new OTP record with fresh timestamp
     new_otp = OTP(
         phone=phone,
         otp=otp_code,
@@ -121,13 +122,13 @@ def send_otp_handler():
         logging.error("Failed to create OTP: %s", e, exc_info=True)
         return internal_error_response()
 
-    # ✅ Send via Twilio or mock
     logging.info("[DEBUG] OTP for %s is %s", phone, otp_code)
 
     return jsonify({"status": "success", "message": "OTP sent"}), 200
 
 # --- Verify OTP ---
 
+@auth_bp.route("/verify-otp", methods=["POST"])
 @limiter.limit(
     lambda: current_app.config["LOGIN_LIMIT_PER_IP"],
     key_func=get_remote_address,
@@ -141,11 +142,9 @@ def verify_otp_handler():
     if not phone or not otp:
         return jsonify({"status": "error", "message": "Phone and OTP are required"}), 400
 
-    # ✅ Look for matching OTP record
     otp_record = OTP.query.filter_by(phone=phone, otp=otp, is_used=False).first()
 
     if not otp_record:
-        # 🔍 Debug info if OTP failed
         recent_otp = OTP.query.filter_by(phone=phone).order_by(OTP.created_at.desc()).first()
         if recent_otp:
             logging.warning("OTP mismatch: submitted=%s, expected=%s", otp, recent_otp.otp)
@@ -154,29 +153,23 @@ def verify_otp_handler():
             logging.warning("No OTP record found for phone: %s", phone)
         return jsonify({"status": "error", "message": "Invalid or expired OTP"}), 401
 
-    # ✅ Check OTP expiry
     otp_expiry_minutes = 10
     if datetime.utcnow() - otp_record.created_at > timedelta(minutes=otp_expiry_minutes):
         return jsonify({"status": "error", "message": "OTP expired"}), 401
 
-    # ✅ Mark OTP as used
     otp_record.is_used = True
 
-    # ✅ Get device/user-agent info safely
     user_agent = request.headers.get("User-Agent", "")[:200]
 
-    # ✅ Create or update UserProfile
     user = UserProfile.query.filter_by(phone=phone).first()
     if not user:
         user = UserProfile(phone=phone)
     user.device_info = user_agent
 
-    # ✅ Generate tokens
     access_token = create_access_token(phone, user.role or "")
     refresh_token = create_refresh_token(phone)
     otp_record.token = "issued"
 
-    # ✅ Commit to DB
     db.session.add(otp_record)
     db.session.add(user)
     try:
